utils: Normalize slashes in multilevel workspace members on Windows
Multilevel cargo workspace members are always written with a forward
slash, for example:

    [workspace]
    members = [
        "apps/android-app",
    ]

On Windows forward slashes are not accepted at all, not even
fs::canonicalize() is able to normalize them. Calling that function or
trying to read a file from a path containing this string results in the
following error:

    Error: The filename, directory name, or volume label syntax is
    incorrect. (os error 123)

Splitting the member name on a forward slash and constructing a new
PathBuf from the parts seems the easiest to get a properly normalized
path with the right separator for the platform.