utils: Strip UNC from canonicalized path which is not supported by glob
We recently taught `fn member()` to look for Cargo.toml files in
workspace members (globs) relative to the workspace root (the directory
containing Cargo.toml with the `[workspace]` listing of these members).
This means the path to the workspace root is now concatenated to the
glob path before searching, which didn't happen before.  That path
happens to be canonicalized just at the top of `fn find_package()`
consequently the whole thing falls apart when Windows UNC paths are used
in places they aren't supported, like `glob`.  No paths are returned,
the package isn't found (`Error::ManifestNotFound` below) and
`cargo-apk` is once again broken on Windows.

We have already been discussing to add `dunce`, now is the right time to
get rid of UNC paths in `cargo-submodule` once and for all.

Fixes: 60fdf8b ("utils: Search for workspace members relative to root manifest (#3)")