Propagate `.cargo/config.toml` `[env]` settings to the process environment
Leaving the caller to read and merge config environment values with the
process environment is not only cumbersome as seen in
[rust-windowing/android-ndk-rs#233], but these values need to be
propagated into the process environment anyway to be usable by any
subprocess spawned by the caller (ie. `cargo-apk` spawning `cargo
rustc`, which can internally spawn a bunch of compilers and linkers
too).

Instead, do this automatically when the caller invokes
`Subcommand::new` while still leaving the door open for crates to read
the environment from the public `subcommand.manifest.env` member.

[rust-windowing/android-ndk-rs#233]: https://github.com/rust-windowing/android-ndk-rs/pull/233