Fix `--manifest-path` to point to `Cargo.toml` instead of a directory
Currently `cargo-subcommand` expects this argument to be a path to the
directory containing project(s), whereas it should actually point to a
file specifically named `Cargo.toml`.

Note that this path doesn't appear to be very strict in `cargo`: even if
pointing to a manifest file within a workspace, packages in the
surrounding workspace are still built, and are reachable through `-p`.

However, there's a strict requirement on the specified manifest to be
part of the workspace, which is not enforced in `cargo-subcommand` yet.