manifest: Skip non-directories when resolving workspace `members`
If globs provided in the `members` array resolve to files, even if
none of its results are (symlinks to) directories or if the path is a
relative path to a file without any globbing directives, `cargo` skips
it without error.  Cargo only errors out if the glob (or again relative
path) didn't resolve to any files at all (which is what `glob::glob()`
should also be doing for us).

It does however support symlinks, meaning we need to canonicalize
every result to follow any symlinks before deducing if the entry is a
directory that needs to be treated as a workspace member crate.